refactor: move project under tasks
